Remove `__pycache__` and add it to .gitignore
Also ignore all `*.pyc`, `*.pyo`, and `*.pyd` files, as these are compiled and
optimized files and libraries that don't need to be in version control, and are
re-created as needed by the Python interpreter dynamically, and depend on the
Python version.

No functional changes.

For more info, see:

* `__pycache__`: https://peps.python.org/pep-3147/
* `*.pyc`: compiled Python modules: https://peps.python.org/pep-0488/
* `*.pyo`: optimized Python modules: https://peps.python.org/pep-0488/
* `*.pyd`: dynamic Python libraries (Windows)
  https://docs.python.org/3/faq/windows.html#is-a-pyd-file-the-same-as-a-dll

Technically, `*.pyo` are obsolete with recent Python versions, but doesn't hurt
to ignore them, as they could still be re-created if someone were to run the
code with older versions of Python.