Add automatic type checking with Mypy and PyType
The type checking currently fails, but we marked it with `continue-on-error:
true` so that it "passes" to avoid breaking the PR review and merge flow.

That said, these type errors should be fixed and `continue-on-error: true`
should be removed to make it obvious when type checking is failing, as it could
be masking real errors if types are not matching or incorrect.