Uncomment `torch` and `numpy`; add `wandb`
All of these modules are directly used in the code, so anyone using them in a
virtual environment would need to install them, and this is impacting the type
checking tests as these libraries are missing, which is throwing errors during
testing.

Also added spaces around `==` for readability as that's allowed in this file.